refactor: consolidate user-related entities, services, and controllers into a dedicated `users` package.
