+++ backend/src/main/java/com/swipelab/auth/application/AuthenticationService.java
@@ -5,8 +5,8 @@ import com.swipelab.dto.request.ResetPasswordRequest;
 import com.swipelab.dto.response.AuthResponse;
 import com.swipelab.exception.EmailVerificationException;
 import com.swipelab.exception.PasswordResetException;
-import com.swipelab.model.entity.User;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.domain.User;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
