+++ backend/src/main/java/com/swipelab/auth/application/JwtService.java
@@ -3,8 +3,8 @@ package com.swipelab.auth.application;
 import com.swipelab.auth.infrastructure.JwtConfig;
 import com.swipelab.dto.response.AuthResponse;
 import com.swipelab.exception.UnauthorizedException;
-import com.swipelab.model.entity.User;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.domain.User;
+import com.swipelab.users.infrastructure.UserRepository;
 import com.swipelab.auth.infrastructure.JwtTokenProvider;
 import com.swipelab.auth.infrastructure.enums.TokenType;
 import lombok.RequiredArgsConstructor;
