+++ backend/src/main/java/com/swipelab/auth/application/OAuth2Service.java
@@ -1,9 +1,9 @@
 package com.swipelab.auth.application;
 
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.model.enums.AuthProvider;
 import com.swipelab.model.enums.UserRole;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
