+++ backend/src/main/java/com/swipelab/classification/application/ClassificationService.java
@@ -7,16 +7,16 @@ import com.swipelab.exception.ResourceNotFoundException;
 import com.swipelab.classification.domain.Classification;
 import com.swipelab.classification.domain.Image;
 import com.swipelab.classification.domain.Label;
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.classification.infrastructure.ClassificationRepository;
 import com.swipelab.classification.infrastructure.ImageRepository;
 import com.swipelab.classification.infrastructure.LabelRepository;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import com.swipelab.classification.application.FraudDetectionService;
 import com.swipelab.gamification.application.BadgeService;
 import com.swipelab.gamification.application.PointsService;
 import com.swipelab.gamification.application.StreakService;
-import com.swipelab.service.user.CredibilityService;
+import com.swipelab.users.application.CredibilityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
