+++ backend/src/main/java/com/swipelab/gamification/application/BadgeService.java
@@ -1,9 +1,9 @@
 package com.swipelab.gamification.application;
 
 import com.swipelab.gamification.domain.Badge;
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.gamification.infrastructure.BadgeRepository;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
