+++ backend/src/main/java/com/swipelab/gamification/application/LeaderboardService.java
@@ -1,7 +1,7 @@
 package com.swipelab.gamification.application;
 
-import com.swipelab.model.entity.User;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.domain.User;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
