+++ backend/src/main/java/com/swipelab/gamification/domain/Leaderboard.java
@@ -7,7 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 
 import java.time.LocalDateTime;
