--- backend/src/main/java/com/swipelab/model/entity/UserStats.java
@@ -1,5 +1,0 @@
-package com.swipelab.model.entity;
-
-public class UserStats {
-    
-}
