+++ backend/src/main/java/com/swipelab/service/analytics/AnalyticsService.java
@@ -6,11 +6,11 @@ import com.swipelab.exception.ResourceNotFoundException;
 import com.swipelab.classification.domain.Classification;
 import com.swipelab.classification.domain.Image;
 import com.swipelab.tasks.domain.Task;
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.classification.infrastructure.ClassificationRepository;
 import com.swipelab.classification.infrastructure.ImageRepository;
 import com.swipelab.tasks.infrastructure.TaskRepository;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
