--- backend/src/main/java/com/swipelab/service/user/ProfileService.java
@@ -1,5 +1,0 @@
-package com.swipelab.service.user;
-
-public class ProfileService {
-    
-}
