+++ backend/src/main/java/com/swipelab/tasks/application/RecipientGroupService.java
@@ -5,8 +5,8 @@ import com.swipelab.dto.request.UpdateRecipientGroupRequest;
 import com.swipelab.dto.response.RecipientGroupResponse;
 import com.swipelab.exception.ResourceNotFoundException;
 import com.swipelab.tasks.domain.RecipientGroup;
-import com.swipelab.model.entity.User;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.domain.User;
+import com.swipelab.users.infrastructure.UserRepository;
 import com.swipelab.tasks.infrastructure.RecipientGroupRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
