+++ backend/src/main/java/com/swipelab/tasks/application/TaskService.java
@@ -6,10 +6,10 @@ import com.swipelab.dto.response.TaskResponse;
 import com.swipelab.exception.ResourceNotFoundException;
 import com.swipelab.mapper.TaskMapper;
 import com.swipelab.tasks.domain.Task;
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.tasks.domain.TaskStatus;
 import com.swipelab.tasks.infrastructure.TaskRepository;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
