+++ backend/src/main/java/com/swipelab/users/application/CredibilityService.java
@@ -1,11 +1,11 @@
-package com.swipelab.service.user;
+package com.swipelab.users.application;
 
 import com.swipelab.classification.domain.Classification;
 import com.swipelab.classification.domain.Label;
-import com.swipelab.model.entity.User;
+import com.swipelab.users.domain.User;
 import com.swipelab.model.enums.UserRole;
 import com.swipelab.classification.infrastructure.ClassificationRepository;
-import com.swipelab.repository.UserRepository;
+import com.swipelab.users.infrastructure.UserRepository;
 import com.swipelab.util.CredibilityCalculator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
