+++ backend/src/main/java/com/swipelab/users/application/ProfileService.java
@@ -1,0 +1,5 @@
+package com.swipelab.users.application;
+
+public class ProfileService {
+    
+}
