+++ backend/src/main/java/com/swipelab/users/domain/UserStats.java
@@ -1,0 +1,5 @@
+package com.swipelab.users.domain;
+
+public class UserStats {
+    
+}
